refactor(builder): extrahiere serp.rs aus single/mod.rs
- build_serp_presentation, build_page_health_presentation → single/serp.rs
